requirements.txt
fastapi
uvicorn
dnspython
python-whois
requests
httpx
scikit-learn
xgboost
joblib
numpy
pandas